+++ wayland-scanner/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+#### Breaking changes
+- Generate tuple struct for enums, and don't use `WEnum`
+
 ## 0.31.8 -- 2025-12-30
 
 - Update `quick-xml` to 0.38